refactor: Add repository base (#9)
* feat: add BaseCsvRepository

Add BaseCsvRepository as an abstract class for repository that uses csv
as its database.

The class provides two methods, enter_reader and enter_writer, which
yields context managers that has been 'entered' so that the concrete
class can do  CRUD operations on the database directly with it.

* refactor: extend BaseCsvRepository on TransactionRepository

* chore: Ruff and isort formatting

* feat: improve BaseCsvRepository

Fix enter_writer overwriting the csv file.

Add method to check if csv file exists, and create if it does not exist.

* refactor: rename Account attribute

Rename 'balance' to 'initial_balance'.

* chore: add .csv templates

* chore: Ruff and isort formatting